requirements.txt
pygame
seaborn
svg.py